Do not reinstall fpm-cookery if present

+++ provision/03-install_fpm-cookery.sh
@@ -1,3 +1,7 @@
 #!/bin/bash
 
-/opt/fpm-refinery/embedded/bin/gem install fpm-cookery --no-ri --no-rdoc
+if [ ! -f "${omnibus_prefix}/embedded/bin/fpm-cook" ]
+then
+    /opt/fpm-refinery/embedded/bin/gem install fpm-cookery --no-ri --no-rdoc
+fi
+
